+++ packages/analytics_hub_firebase/example/main.dart
@@ -1,14 +1,6 @@
 import 'package:analytics_hub/analytics_hub.dart';
 import 'package:analytics_hub_firebase/analytics_hub_firebase.dart';
 
-class EmptySessionDelegate implements HubSessionDelegate {
-  @override
-  Stream<Session?> get sessionStream => Stream.value(null);
-
-  @override
-  Future<Session?> getSession() async => null;
-}
-
 class ExampleEvent extends Event {
   const ExampleEvent({required this.creativeName}) : super('select_promotion');
 
@@ -25,7 +17,6 @@ class ExampleEvent extends Event {
 
 Future<void> main() async {
   final hub = AnalyticsHub(
-    sessionDelegate: EmptySessionDelegate(),
     providers: [
       FirebaseAnalyticsHubProvider.fromInstance(),
     ],