+++ packages/analytics_hub_firebase/lib/analytics_hub_firebase.dart
@@ -9,7 +9,6 @@
 /// final provider = FirebaseAnalyticsHubProvider.fromInstance();
 /// final hub = AnalyticsHub(
 ///   providers: [provider],
-///   sessionDelegate: mySessionDelegate,
 /// );
 /// await hub.initialize();
 /// ```