+++ packages/analytics_hub_mixpanel/example/main.dart
@@ -2,14 +2,6 @@ import 'package:analytics_hub/analytics_hub.dart';
 import 'package:analytics_hub_mixpanel/analytics_hub_mixpanel.dart';
 import 'package:mixpanel_flutter/mixpanel_flutter.dart';
 
-class EmptySessionDelegate implements HubSessionDelegate {
-  @override
-  Stream<Session?> get sessionStream => Stream.value(null);
-
-  @override
-  Future<Session?> getSession() async => null;
-}
-
 class ExampleEvent extends Event {
   const ExampleEvent({required this.exampleProperty})
       : super('example_log_event');
@@ -32,7 +24,6 @@ Future<void> main() async {
   );
 
   final hub = AnalyticsHub(
-    sessionDelegate: EmptySessionDelegate(),
     providers: [
       MixpanelAnalyticsHubProvider(mixpanel: mixpanel),
     ],